% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_num_anteriores.R
\name{suma_num_anteriores}
\alias{suma_num_anteriores}
\title{Suma anteriores}
\usage{
suma_num_anteriores(a, b)
}
\arguments{
\item{a}{a numeric value}

\item{b}{a numeric value}
}
\value{
the sum of the previus input
}
\description{
Suma de los numeros anteriores a los dos numeros presentes
}

