% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_oso.R
\name{test.oso}
\alias{test.oso}
\title{Test de osos}
\usage{
test.oso(x = NULL, y = NULL, z = NULL)
}
\arguments{
\item{x}{a character value}

\item{y}{a numeric value}

\item{z}{a character value}
}
\value{
Resultado del test
}
\description{
Se evaluará el conocimiento sobre los osos
}

